--- README.md
@@ -1,11 +1,0 @@
-### Warning This repo is being rewritten in typescript for a faster and better development experience
-# Density
-### What is it?
-Density is a web framework for drawing simple shapes and sprites for any project. But Density can be very powerful.
-
-### Getting Started
-As Density is still a WIP project it is recommended to download/clone the src folder for a quick start. You can use the density-builder tool (alpha) to minimize all of the JS in your project and make a production build.
-
-### TODO
-- Webpack support
-- Update cluster renderer for 2.0
