--- src/assets/assets.json
@@ -1,1 +1,0 @@
-["donut.png"]
