--- src/density.css
@@ -1,66 +1,0 @@
-body {
-  margin: 0px;
-  padding: 0px;
-  overflow: hidden;
-}
-
-canvas {
-  margin: 0%;
-  width: 100%;
-  height: 100%;
-}
-
-#loadingScreen {
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  text-align: center;
-  background-color: black;
-  z-index: 20;
-}
-
-#densityLogoContainer {
-  height: 100%;
-  /* Can be anything */
-  width: 100%;
-  /* Can be anything */
-  position: relative;
-}
-
-#densityLogo {
-  max-height: 50%;
-  max-width: 100%;
-  width: auto;
-  height: auto;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: auto;
-  animation-name: logoAnimation;
-  animation-duration: 1s;
-}
-
-@keyframes logoAnimation {
-  from {
-    max-height: 40%;
-    max-width: 80%;
-    opacity: 0%;
-  }
-
-  to {
-    max-height: 50%;
-    max-width: 100%;
-    opacity: 100%;
-  }
-}
-
-iframe {
-  height: 400px;
-  width: 350px;
-  border: none;
-  border-radius: 15px;
-}
