--- src/density.json
@@ -1,1 +1,0 @@
-{ "buildSettings": { "minify": true } }
