--- src/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"
-    />
-    <title>Density Engine | 2.0</title>
-    <link rel="stylesheet" href="density.css" type="text/css" />
-    <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-  </head>
-
-  <body>
-    <script src="game.js" type="module"></script>
-    <div id="loadingScreen">
-      <div id="densityLogoContainer">
-        <img src="Density-Logo.svg" id="densityLogo" />
-      </div>
-    </div>
-    <canvas id="maincanvas" width="0" height="0"></canvas>
-    <!--Please note that this has been replaced by the new <script src="Density-Initializer.js"></script>-->
-  </body>
-</html>
